+++ docs/CHANGELOG.md
@@ -14,6 +14,14 @@ nav_order: 5
 
     *Erinna Chen*
 
+* Add `with_rendered_component_path` helper for writing component system tests.
+
+    *Edwin Mak*
+
+* Include gem name and deprecation horizon in every deprecation message.
+
+    *Jan Klimo*
+
 ## 2.78.0
 
 * Support variants with dots in their names.